:bug: Fix bugs originating from #52e470a
