+++ platforms/platform-alexa/src/AlexaPlatform.ts
@@ -11,6 +11,7 @@ import { AlexaOutputTemplateConverterStrategy, AlexaResponse } from '@jovotech/o
 import { AlexaRequest } from './AlexaRequest';
 import { Alexa } from './Alexa';
 import { AlexaUser } from './AlexaUser';
+import { SUPPORTED_APL_ARGUMENT_TYPES } from './constants';
 
 export interface AlexaConfig extends ExtensibleConfig {
   output: {
@@ -54,7 +55,7 @@ export class AlexaPlatform extends Platform<AlexaRequest, AlexaResponse, Alexa, 
   }
 
   private beforeRequest = (handleRequest: HandleRequest, jovo: Jovo) => {
-    if (!(jovo.$platform instanceof AlexaPlatform)) {
+    if (!(jovo.$platform instanceof Alexa)) {
       return;
     }
     // Generate generic output to APL if supported and set in config
@@ -64,15 +65,18 @@ export class AlexaPlatform extends Platform<AlexaRequest, AlexaResponse, Alexa, 
     if (jovo.$alexa?.$request?.request?.type === 'Alexa.Presentation.APL.UserEvent') {
       const requestArguments = jovo.$alexa.$request.request.arguments || [];
       requestArguments.forEach((argument) => {
-        if (typeof argument === 'object' && argument.type === 'QuickReply') {
+        if (typeof argument === 'object' && SUPPORTED_APL_ARGUMENT_TYPES.includes(argument?.type)) {
           if (argument.intent) {
             jovo.$nlu.intent = { name: argument.intent };
           }
           if (argument.entities) {
-            const entityMap: EntityMap = {};
-            argument.entities.forEach((entity: Entity) => {
-              entityMap[entity.name!] = entity;
-            });
+            const entityMap: EntityMap = argument.entities.reduce(
+              (entityMap: EntityMap, entity: Entity) => {
+                entityMap[entity.name] = entity;
+                return entityMap;
+              },
+              {},
+            );
             jovo.$nlu.entities = { ...entityMap };
             jovo.$entities = entityMap;
           }
